En attendant de faire mieux, il faut caster lors de la soustraction
git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@16485 ee107e0b-c2fa-0310-842a-ec2723fb34f4

+++ tbx/include/tbx_macros.h
@@ -512,7 +512,7 @@ do { \
  * variables (the new value is directly given to other caches)
  */
 #if defined(MARCEL) && defined(X86_64_ARCH)
-#define TBX_SHARED_SET(v,n) asm("lock add %1,%0":"=m"(v):"r"(n-v))
+#define TBX_SHARED_SET(v,n) asm("lock add %1,%0":"=m"(v):"r"((unsigned long)n-(unsigned long)v))
 #else
 #define TBX_SHARED_SET(v,n) (v) = (n);
 #endif
